refactor: use workspace, add benchmark-lib crate
