+++ benchmark-lib/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "benchmark-lib"
+version = "0.1.0"
+license.workspace = true
+edition.workspace = true
+
+[dependencies]
+bevy_ecs = "0.13.2"
+bytemuck = { version = "1.15.0", features = [ "derive" ] }
+criterion = { version = "0.5.1", features = ["html_reports"] }
+specs = "0.20.0"
+xanadu = { path = "../" }
+
+[[bench]]
+name = "bench"
+harness = false
+
